zkDSL compiler: separate compile-time evaluation from lowering
lower.rs was 3,123 lines with no internal boundary. The clearest one available
is already latent in the code: twelve methods answer "what is this expression
worth before anything runs", and every one of them takes &self and emits
nothing. That is what lets a caller ask the question without paying for the
answer, and it is why an index that folds costs no instruction.

They move to lower/eval.rs, where the module doc states the rule and &self makes
it structural: a function there cannot emit, because emitting needs &mut self.
lower.rs drops to 2,884 lines.

The module doc also states the thing the split makes visible: there are two
answers, not one, and the position of a use decides which is wanted.
try_const_int reads an expression as an integer, which is what a size, an index,
a bound and an exponent want; try_field_const reads the same expression as a
field element, where `+` is XOR, which is what a value wants. They disagree on
any sum of overlapping integers, and the language cannot yet say which was
meant.

A pure move: no digest changes, the guest is unchanged.
